+++ controllers/componentsController.js
@@ -3,7 +3,6 @@ const { prisma } = require("../db/prisma");
 
 const componentsController = {
   getFileDetails: asyncHandler(async (req, res) => {
-    console.log("getFileDetails running...");
     const { fileID } = req.params;
     const file = await prisma.file.findFirst({
       where: {
@@ -14,16 +13,12 @@ const componentsController = {
     res.render("fileDetails", { file });
   }),
   getUploadFileForm: asyncHandler(async (req, res) => {
-    console.log("getUploadFileForm running...");
     res.render("uploadForm");
   }),
   getAddFolderForm: asyncHandler(async (req, res) => {
-    // folderName_onInput
-    console.log("getAddFolderForm running...");
     res.render("createFolderForm");
   }),
   getEditFileForm: asyncHandler(async (req, res) => {
-    console.log("getEditFileForm");
     const { fileID } = req.params;
     const file = await prisma.file.findFirst({
       where: {
@@ -36,7 +31,6 @@ const componentsController = {
     });
   }),
   getEditFolderForm: asyncHandler(async (req, res) => {
-    console.log("getEditFolderForm");
     const { folderID } = req.params;
     const folder = await prisma.folder.findFirst({
       where: {
@@ -49,7 +43,6 @@ const componentsController = {
     });
   }),
   getDeleteFileForm: asyncHandler(async (req, res) => {
-    console.log("getDeleteFileForm running...");
     const { fileID } = req.params;
     const file = await prisma.file.findFirst({
       where: {
@@ -63,7 +56,6 @@ const componentsController = {
     });
   }),
   getDeleteFolderForm: asyncHandler(async (req, res) => {
-    console.log("getDeleteFolderForm running...");
     const { folderID } = req.params;
     const folder = await prisma.folder.findFirst({
       where: {
@@ -76,7 +68,6 @@ const componentsController = {
     });
   }),
   getShareFolderForm: asyncHandler(async (req, res) => {
-    console.log("getShareFolderForm running...");
     const { folderID } = req.params;
 
     res.render("shareFolderForm", { folderID });