+++ controllers/indexController.js
@@ -2,7 +2,6 @@ const asyncHandler = require("express-async-handler");
 
 const indexController = {
   getIndex: asyncHandler((req, res) => {
-    console.log("getIndex running...");
     res.render("index");
   }),
 };