+++ events/dialogController.js
@@ -2,12 +2,8 @@ const dialog = document.querySelector("dialog");
 
 const openDialog = async (e) => {
   const { url, formAction } = e.currentTarget.dataset;
-  console.log("e.currentTarget:", e.currentTarget);
-  console.log("e.currentTarget.dataset:", e.currentTarget.dataset);
-  console.log("url:", url);
   const content = await fetchContent(url, formAction);
-  console.log(content);
-  console.log("window.location:", window.location);
+
   // Need to set the form's action attribute
   // /drive/folder/:folderID/files/upload
   // /drive/files/upload
@@ -18,18 +14,15 @@ const openDialog = async (e) => {
 };
 
 const closeDialog = (e) => {
-  console.log("closeDialog running...");
   dialog.close();
 };
 
 const onCloseHandler = () => {
-  console.log("onCloseHandler running...");
   dialog.firstChild.remove();
 };
 
 // stopPropagation on the dialog element's child/children element(s)
 const stopPropagation = (e) => {
-  console.log("stopPropagation running...");
   e.stopPropagation();
 };
 
@@ -39,23 +32,20 @@ const fetchContent = async (url, formAction) => {
     method: "GET",
   })
     .then((res) => {
+      if (res.status === 404) throw new Error("Component does not exist");
       return res.text();
     })
     .then((html) => {
       const parser = new DOMParser();
       const doc = parser.parseFromString(html, "text/html");
-      console.log("--------------------------------------------");
-      console.log(doc);
+
+      const htmlContent = doc.body.firstElementChild;
       if (formAction) {
         // What are the advantages and disadvantages using doc.body.firstChild vs querySelector()
-        // const form = doc.body.firstChild;
-        const form = doc.querySelector("form");
-        form.setAttribute("action", formAction);
-        return form;
+        htmlContent.setAttribute("action", formAction);
       }
-      console.log("doc.body:", doc.body.firstElementChild);
 
-      return doc.querySelector("article");
+      return htmlContent;
     })
     .catch((err) => {
       // Is there a proper way to handle errors when fetching components?