+++ validators/uploadValidator.js
@@ -7,53 +7,26 @@ const {
 const allowedMimeTypes = require("./allowedMimeTypes/allowedMimeTypes");
 
 const validateFileExtension = (ext) => {
-  console.log("validateFileExtension running...");
-  console.log("ext:", ext);
   if (!allowedMimeTypes[ext])
     throw new Error(`File extension, "${ext}", is not supported.`);
 };
 
 const validateFileMIMEType = (ext, mimetype) => {
-  console.log("validateFileMIMEType running...");
-  // const type = mimetype.split("/", 2);
   const mimetypeRegex = new RegExp(`^${allowedMimeTypes[ext]}$`);
   const regexResult = mimetypeRegex.test(mimetype);
   if (!regexResult)
     throw new Error(`MIME Type, "${mimetype}", is not supported.`);
-  /*  switch (type) {
-    case "application":
-      // ^((msword|ogg|pdf)$)|(vnd.((oasis|openxmlformats-officedocument)(.(opendocument|presentationml|wordprocessingml|spreadsheetml).(presentation|spreadsheet|text|document|sheet))|(ms-(excel|powerpoint))))$
-      break;
-    case "audio":
-      // ^(audio\/(aac|midi|x-midi|mpeg|ogg|wav))$
-      break;
-    case "image":
-      // ^(image\/(bmp|gif|jpeg|png|svg\+xml))$
-      break;
-    case "text":
-      // ^(text\/(css|csv|html|plain))$
-      break;
-    case "video":
-      // ^(video\/(x-msvideo|mp4|mpeg|ogg))$
-      break;
-    default:
-  } */
 };
 
 const validateFileSize = (filename, bytes) => {
-  console.log("validateFileSize running...");
-  // const fileSizeLimit = 1;
   const fileSizeLimit = 5000000;
   if (bytes > fileSizeLimit)
     throw `The file, ${filename}, exceeds file size limit.`;
 };
 
 const validateFiles = async (values, { req }) => {
-  console.log("validateFiles running...");
-  console.log("req.files:", req.files);
   const filesLimit = 2;
   const { files } = req;
-  console.log("files:", files);
 
   // Is this needed if submit button is disabled?
   if (files.length === 0) throw new Error("No files selected.");
@@ -83,12 +56,9 @@ const filesSchema = {
 
 const validateUpload = (view) => {
   return asyncHandler(async (req, res, next) => {
-    console.log("validateUpload...");
-    console.log("req.originalUrl:", req.originalUrl);
     await checkSchema(filesSchema, ["body"]).run(req);
     const errors = validationResult(req);
     const inputs = matchedData(req, { onlyValidData: false });
-    console.log("errors:", errors);
     if (!errors.isEmpty()) {
       return res.status(422).render(view, {
         errors: errors.mapped(),